requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
google-genai>=1.5.0
sse-starlette>=2.0.0
langgraph>=1.0.0
langchain-core>=1.2.0
